#161: Added support for UPI CLI options
