+++ common.js
@@ -43,6 +43,8 @@ const CURRENT_DATA_DICTIONARY_VERSION = DATA_DICTIONARY_VERSIONS.v2_0,
   PREVIOUS_WEB_API_CORE_VERSION = WEB_API_CORE_VERSIONS.v2_0_0,
   DEFAULT_WEB_API_CORE_VERSION = WEB_API_CORE_VERSIONS.v2_0_0;
 
+const DEFAULT_UPI_VERSION = '2.0';
+
 const COMMANDER_LOG_FILE_NAME = 'commander.log';
 
 const EMPTY_STRING = '',
@@ -578,6 +580,7 @@ const readZipFileContents = path => {
 module.exports = {
   NOT_OK,
   DEFAULT_DD_VERSION,
+  DEFAULT_UPI_VERSION,
   DEFAULT_WEB_API_CORE_VERSION,
   DATA_DICTIONARY_VERSIONS,
   WEB_API_CORE_VERSIONS,
