+++ lib/certification/upi/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  runUpiTests: () => {}
+};
