+++ lib/misc/index.js
@@ -1,6 +1,9 @@
 'use strict';
 
-const { pascalCase } = require('pascal-case');
+/* gross workaround for ESM */
+let changeCase;
+import("change-case").then(module => changeCase = module);
+
 const { getReferenceMetadata } = require('@reso/reso-certification-etl');
 const { writeFile } = require('fs/promises');
 
@@ -13,7 +16,7 @@ const generatePascalCaseODataValuesFromReferenceMetadata = async (version = '1.7
     resources: resources.flatMap(({ resourceName }) => {
       if (!resourceName) return [];
 
-      const processedResourceName = pascalCase(resourceName);
+      const processedResourceName = changeCase.pascalCase(resourceName);
 
       if (processedResourceName !== resourceName) {
         return {
@@ -27,8 +30,8 @@ const generatePascalCaseODataValuesFromReferenceMetadata = async (version = '1.7
     fields: fields.flatMap(({ resourceName, fieldName }) => {
       if (!(resourceName && fieldName)) return [];
 
-      const processedResourceName = pascalCase(resourceName),
-        processedFieldName = pascalCase(fieldName);
+      const processedResourceName = changeCase.pascalCase(resourceName),
+        processedFieldName = changeCase.pascalCase(fieldName);
 
       if (processedFieldName !== fieldName) {
         return {
@@ -62,8 +65,8 @@ const generatePascalCaseODataValuesFromReferenceMetadata = async (version = '1.7
 
       if (!(lookupName && lookupValue)) return [];
 
-      const processedLookupName = pascalCase(lookupName?.substring(lookupName.lastIndexOf('.') + 1)),
-        processedLegacyODataValue = pascalCase(lookupValue);
+      const processedLookupName = changeCase.pascalCase(lookupName?.substring(lookupName.lastIndexOf('.') + 1)),
+        processedLegacyODataValue = changeCase.pascalCase(lookupValue);
 
       if (processedLegacyODataValue !== legacyODataValue) {
         return {
