+++ package.json
@@ -35,6 +35,7 @@
     "aws-sdk": "^2.1354.0",
     "axios": "^0.28.0",
     "chalk": "^4.1.2",
+    "change-case": "^5.4.4",
     "commander": "^9.4.0",
     "dotenv": "^16.0.1",
     "fast-xml-parser": "^4.3.1",
@@ -42,7 +43,6 @@
     "fs-extra": "^10.0.0",
     "humanize-duration": "^3.30.0",
     "mocha": "^9.1.3",
-    "pascal-case": "^3.1.2",
     "yargs": "^16.2.0",
     "yauzl": "^3.1.1"
   }
